+++ src/Container/NotFoundException.php
@@ -20,9 +20,11 @@
 
 namespace YoutubeDownloader\Container;
 
+use Psr\Container\NotFoundExceptionInterface;
+
 /**
  * No entry was found in the container.
  */
-class NotFoundException extends ContainerException
+class NotFoundException extends ContainerException implements NotFoundExceptionInterface
 {
 }